+++ SnakeController/World.cpp
@@ -1,0 +1,7 @@
+#pragma once
+#include "World.hpp"
+
+bool World::isPositionOutsideMap(int x, int y) const
+{
+    return x < 0 or y < 0 or x >= m_mapDimension.first or y >= m_mapDimension.second;
+}
